+++ src/OpenTelemetry.Exporter.Console/ConsoleMetricExporter.cs
@@ -146,8 +146,21 @@ namespace OpenTelemetry.Exporter
                         }
                         else
                         {
-                            // TODO: Consider how/if to display buckets for exponential histograms.
-                            bucketsBuilder.AppendLine("Buckets are not displayed for exponential histograms.");
+                            var exponentialHistogramData = metricPoint.GetExponentialHistogramData();
+                            var scale = exponentialHistogramData.Scale;
+
+                            if (exponentialHistogramData.ZeroCount != 0)
+                            {
+                                bucketsBuilder.AppendLine($"Zero Bucket:{exponentialHistogramData.ZeroCount}");
+                            }
+
+                            var offset = exponentialHistogramData.PositiveBuckets.Offset;
+                            foreach (var bucketCount in exponentialHistogramData.PositiveBuckets)
+                            {
+                                var lowerBound = Base2ExponentialBucketHistogram.LowerBoundary(offset, scale).ToString(CultureInfo.InvariantCulture);
+                                var upperBound = Base2ExponentialBucketHistogram.LowerBoundary(++offset, scale).ToString(CultureInfo.InvariantCulture);
+                                bucketsBuilder.AppendLine($"({lowerBound}, {upperBound}]:{bucketCount}");
+                            }
                         }
 
                         valueDisplay = bucketsBuilder.ToString();
@@ -176,6 +189,7 @@ namespace OpenTelemetry.Exporter
                     }
 
                     var exemplarString = new StringBuilder();
+                    /* Commenting out as Exemplars is marked internal
                     foreach (var exemplar in metricPoint.GetExemplars())
                     {
                         if (exemplar.Timestamp != default)
@@ -206,6 +220,7 @@ namespace OpenTelemetry.Exporter
                             exemplarString.AppendLine();
                         }
                     }
+                    */
 
                     msg = new StringBuilder();
                     msg.Append('(');